Analyze per-field numbers and strings
Extract out JSON field logic to its own file.

For numbers, create a guess of the number's real type by examining
each number. For integers, use the min and max to determine if
unsigned integers work, and the number of bytes required to store.
For floats, examine the fraction bits to see if there are non-zero
entries in the float64-specific fraction bits. If they are all zero
for every number, assume float32.

Note: The float logic does not handle the case of exponents outside
of the float32 range. Examining this is currently unnecessary for
our application, but could be added in fairly easily in the future.

Signed-off-by: Will Beason <willbeason@gmail.com>